# packages/reading-assistant-query/src/api.schemas.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: 0.1.0
 */
export interface AuthResponse {
  email: string;
  user_id: string;
}

/**
 * The response payload sent after successfully creating a session.
 */
export interface CreateSessionResponse {
  document_id: string;
  session_id: string;
  user_id: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface SignupRequest {
  email: string;
  password: string;
}